Fixes bug with embedded tags
If the item.description contains embedded tags, and if the text truncation happens to land in the middle of a tag, the formatting of the resulting description gets broken. For example, if the text contains <a href="something"> and the truncation happens inside that tag, then the tag never closes, and the rest of the page is broken.

The proposes fix strips out all tags, which might not be expected behavior for everyone, but fixes that problem.